Added tests for shift method

+++ pybboxes/types/yolo_bounding_box.py
@@ -38,14 +38,14 @@ class YoloBoundingBox(BaseBoundingBox):
             return x_tl, y_tl, x_br, y_br
         return BoundingBox(x_tl, y_tl, x_br, y_br, image_size=self.image_size, strict=self.strict)
 
-    def shift(self, threshold: Tuple[int, int]) -> "YoloBoundingBox":
+    def shift(self, threshold: Tuple[float, float]) -> "YoloBoundingBox":
         """Returns a new bounding box shifted by the given thresholds. The new
         bounding box has same image shape, and other properties as the current
         object.
 
         Parameters
         ----------
-        threshold: Tuple[int, int]
+        threshold: Tuple[float, float]
             The amount to shift the bounding box. The first value is the
                 amount to shift the x-coordinate, and the second value is the
                 amount to shift the y-coordinate.
